+++ basic_test.sh
@@ -34,5 +34,14 @@ echo "Testing: gl_mirror show-account"
 gl_mirror show-account
 
 # These modify the file system, not safe to run just to see if they blow up.
-# clone-all", "pull-all", "update-from-main", "prune-all"],
+gl_mirror clone-all --dry-run
+gl_mirror pull-all --dry-run
+gl_mirror update-from-main --dry-run
+gl_mirror prune-all --dry-run
+
+gh_mirror clone-all --dry-run
+gh_mirror pull-all --dry-run
+gh_mirror update-from-main --dry-run
+gh_mirror prune-all --dry-run
+
 # "init", "sync-config"