+++ git_mirror/menu.py
@@ -41,6 +41,7 @@ def get_command_info(args: argparse.Namespace) -> None:
         ("Cross-repo report", "cross-repo-report"),
         ("Cross-repo sync", "cross-repo-sync"),
         ("Cross-repo Template Initialization", "cross-repo-init"),
+        ("Poetry ", "cross-repo-init"),
         ("Main Menu", "Main Menu"),
     ]
 
@@ -51,6 +52,7 @@ def get_command_info(args: argparse.Namespace) -> None:
         "PyPI Commands": ["pypi-status", "Main Menu"],
         "Source Control Host Commands": ["show-account", "Main Menu"],
         "Template Sync": ["cross-repo-report", "cross-repo-sync", "cross-repo-init", "Main Menu"],
+        "Poetry Commands": ["poetry-relock", "Main Menu"],
         "Exit": [],
     }