+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.3] - 2024-02-24
+
+### Added
+
+- Started cross-repo sync feature. Beta quality
+- Started poetry dependency lock. Command not surfaced yet.
+
+### Fixed
+
+- more unit tests, found bugs in the config UI
+
 ## [0.3.2] - 2024-02-21
 
 ### Added